# submissions/sama-panda/ui/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Court Time Planner API
servers:
  - url: /api
paths:
...
  /cases/{id}:
    get:
      operationId: getCase
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Case" }
...
components:
  schemas:
...
    Reason:
      type: object
      required: [code, detail]
      properties:
        code: { type: string }
        detail: { type: string }
    Case:
      type: object
      required: [id, filingNumber, filingDate, advocateId, partyId, stage, purpose, lastHearing, totalHearings, ageYears, flags, waitingOn, history, reasons]
      properties:
        id: { type: string }
        filingNumber: { type: string }
        filingDate: { type: string }
        advocateId: { type: string }
        partyId: { type: string }
        stage: { type: string }
        purpose: { type: string }
        lastHearing: { type: string }
        totalHearings: { type: integer }
        ageYears: { type: number }
        flags:
          type: array
          items: { type: string }
        waitingOn: { type: string }
        history:
          type: array
          items: { type: string }
        reasons:
          type: array
          items: { $ref: "#/components/schemas/Reason" }
...